+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+-   Fix a bug in the deprecated callback handlers of Form component. #7356
+
 # 8.0.0
 
 -   Fix commonjs module build, allow package to be built in isolation. #7286